maybe solved unusual button click order bugs

+++ server/routes/operations.js
@@ -1,3 +1,4 @@
+//This file routes requests based on what math operation was requested. The posts update a result property. The get obtains the result object.
 var express = require('express');
 var router = express.Router();
 
@@ -8,29 +9,29 @@ router.get('/', function(req,res){
 })
 
 router.post('/add', function(req,res){
-  var x = parseInt(req.body.x);
-  var y = parseInt(req.body.y);
+  var x = parseFloat(req.body.x);
+  var y = parseFloat(req.body.y);
   result.result = x + y;
   res.sendStatus(200);
 })
 
 router.post('/subtract', function(req,res){
-  var x = parseInt(req.body.x);
-  var y = parseInt(req.body.y);
+  var x = parseFloat(req.body.x);
+  var y = parseFloat(req.body.y);
   result.result = x - y;
   res.sendStatus(200);
 })
-//add other post functions here for the /multiply, /divide, /subtract requests
+
 router.post('/multiply', function(req,res){
-  var x = parseInt(req.body.x);
-  var y = parseInt(req.body.y);
+  var x = parseFloat(req.body.x);
+  var y = parseFloat(req.body.y);
   result.result = x * y;
   res.sendStatus(200);
 })
 
 router.post('/divide', function(req,res){
-  var x = parseInt(req.body.x);
-  var y = parseInt(req.body.y);
+  var x = parseFloat(req.body.x);
+  var y = parseFloat(req.body.y);
   if(y==0){
     res.status(500).send('Divide by zero error');
   } else {
